+++ lib/service/FirebaseMessaging.dart
@@ -19,8 +19,10 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:math';
 import 'dart:ui';
+
+import 'package:firebase_messaging/firebase_messaging.dart' as firebase_messaging;
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:http/http.dart';
-import 'package:firebase_messaging/firebase_messaging.dart' as FirebaseMessagingPlugin;
 import 'package:illinois/model/Health.dart';
 import 'package:illinois/model/UserProfile.dart';
 import 'package:illinois/service/AppLivecycle.dart';
@@ -38,6 +40,8 @@ import 'package:illinois/service/Storage.dart';
 import 'package:illinois/service/UserProfile.dart';
 import 'package:illinois/service/LocalNotifications.dart';
 import 'package:illinois/utils/Utils.dart';
+
+const String _channelId = "Notifications_Channel_ID";
 
 class FirebaseMessaging with Service implements NotificationsListener {
 
@@ -64,6 +68,14 @@ class FirebaseMessaging with Service implements NotificationsListener {
   };
 
 
+
+  final AndroidNotificationChannel _channel = AndroidNotificationChannel(
+    _channelId, // id
+    "Safer Illinois",
+    "Receive notifications",
+    importance: Importance.high,
+  );
+
   String   _token;
   String   _projectID;
   DateTime _pausedDateTime;
@@ -74,8 +86,7 @@ class FirebaseMessaging with Service implements NotificationsListener {
 
   FirebaseMessaging._internal();
   static final FirebaseMessaging _firebase = FirebaseMessaging._internal();
-  FirebaseMessagingPlugin.FirebaseMessaging _firebaseMessaging = FirebaseMessagingPlugin.FirebaseMessaging();
-
+  final FlutterLocalNotificationsPlugin _firebaseMessaging = FlutterLocalNotificationsPlugin();
   factory FirebaseMessaging() {
     return _firebase;
   }
@@ -114,16 +125,31 @@ class FirebaseMessaging with Service implements NotificationsListener {
   Future<void> initService() async {
     
     // Cache messages until UI is displayed
-    _messagesCache = List<Map<String, dynamic>>();
+    _messagesCache = <Map<String, dynamic>>[];
 
-    _firebaseMessaging.configure(
-      onMessage: _onFirebaseMessage,
-      onBackgroundMessage: null, // causes exception in FirebaseMessaging plugin 
-      onLaunch: _onFirebaseLaunch,
-      onResume: _onFirebaseResume,
+    //firebase_messaging.FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
+
+    await _firebaseMessaging.resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()?.createNotificationChannel(_channel);
+
+    await firebase_messaging.FirebaseMessaging.instance.setForegroundNotificationPresentationOptions(
+      alert: true,
+      badge: true,
+      sound: true,
     );
-    
-    _firebaseMessaging.getToken().then((String token) {
+
+    firebase_messaging.FirebaseMessaging.onMessage.listen((firebase_messaging.RemoteMessage message) {
+      _onFirebaseMessage(message.data);
+    });
+
+    firebase_messaging.FirebaseMessaging.onMessageOpenedApp.listen((firebase_messaging.RemoteMessage message) {
+      print('A new onMessageOpenedApp event was published!');
+      _onFirebaseMessage(message.data);
+    });
+
+
+
+
+    firebase_messaging.FirebaseMessaging.instance.getToken().then((String token) {
       _token = token;
       Log.d('FCM: token: $token');
       NotificationService().notify(notifyToken, null);
@@ -276,16 +302,6 @@ class FirebaseMessaging with Service implements NotificationsListener {
 
   Future<dynamic> _onFirebaseMessage(Map<String, dynamic> message) async {
     Log.d("FCM: onFirebaseMessage");
-    _onMessageProcess(message);
-  }
-
-  Future<dynamic> _onFirebaseLaunch(Map<String, dynamic> message) async {
-    Log.d("FCM: onFirebaseLaunch");
-    _onMessageProcess(message);
-  }
-
-  Future<dynamic> _onFirebaseResume(Map<String, dynamic> message) async {
-    Log.d("FCM: onFirebaseResume");
     _onMessageProcess(message);
   }