+++ lib/service/Network.dart
@@ -65,14 +65,8 @@ class Network  {
 
   Future<Http.Response> _get2(dynamic url, { String body, Encoding encoding, Map<String, String> headers, int timeout, Http.Client client }) async {
     try {
-      
-      Uri uri;
-      if (url is Uri) {
-        uri = url;
-      }
-      else if (url is String) {
-        uri = Uri.parse(url);
-      }
+
+      Uri uri = _uriFromUrlString(url);
 
       if (uri != null) {
         
@@ -120,19 +114,21 @@ class Network  {
   Future<Http.Response> _get(url, { String body, Encoding encoding, Map<String, String> headers, int auth, int timeout, Http.Client client} ) async {
     if (Connectivity().isNotOffline) {
       try {
-        if (url != null) {
+        Uri uri = _uriFromUrlString(url);
 
-          Map<String, String> requestHeaders = _prepareHeaders(headers, auth, url);
+        if (uri != null) {
+
+          Map<String, String> requestHeaders = await _prepareHeaders(headers, auth, uri);
           
           Future<Http.Response> response;
           if (body != null) {
-            response = _get2(url, headers: requestHeaders, body: body, encoding: encoding, timeout: timeout, client: client);
+            response = _get2(uri, headers: requestHeaders, body: body, encoding: encoding, timeout: timeout, client: client);
           }
           else if (client != null) {
-            response = client.get(url, headers: requestHeaders);
+            response = client.get(uri, headers: requestHeaders);
           }
           else {
-            response = Http.get(url, headers: requestHeaders);
+            response = Http.get(uri, headers: requestHeaders);
           }
           
           if ((response != null) && (timeout != null)) {
@@ -176,7 +172,8 @@ class Network  {
   Future<Http.Response> _post(url, { body, Encoding encoding, Map<String, String> headers, int auth, int timeout}) async{
     if (Connectivity().isNotOffline) {
       try {
-        Future<Http.Response> response = (url != null) ? Http.post(url, headers: _prepareHeaders(headers, auth, url), body: body, encoding: encoding) : null;
+        Uri uri = _uriFromUrlString(url);
+        Future<Http.Response> response = (uri != null) ? Http.post(uri, headers: await _prepareHeaders(headers, auth, uri), body: body, encoding: encoding) : null;
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout), onTimeout: _responseTimeoutHandler) : response;
       } catch (e) {
         Log.d(e?.toString());
@@ -214,10 +211,11 @@ class Network  {
   Future<Http.Response> _put(url, { body, Encoding encoding, Map<String, String> headers, int auth, int timeout, Http.Client client }) async {
     if (Connectivity().isNotOffline) {
       try {
-        Future<Http.Response> response = (url != null) ?
+        Uri uri = _uriFromUrlString(url);
+        Future<Http.Response> response = (uri != null) ?
           ((client != null) ?
-            client.put(url, headers: _prepareHeaders(headers, auth, url), body: body, encoding: encoding) :
-              Http.put(url, headers: _prepareHeaders(headers, auth, url), body: body, encoding: encoding)) :
+            client.put(uri, headers: await _prepareHeaders(headers, auth, uri), body: body, encoding: encoding) :
+              Http.put(uri, headers: await _prepareHeaders(headers, auth, uri), body: body, encoding: encoding)) :
             null;
 
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout), onTimeout: _responseTimeoutHandler) : response;
@@ -257,7 +255,8 @@ class Network  {
   Future<Http.Response> _patch(url, { body, Encoding encoding, Map<String, String> headers, int auth, int timeout }) async {
     if (Connectivity().isNotOffline) {
       try {
-        Future<Http.Response> response = (url != null) ? Http.patch(url, headers: _prepareHeaders(headers, auth, url), body: body, encoding: encoding) : null;
+        Uri uri = _uriFromUrlString(url);
+        Future<Http.Response> response = (uri != null) ? Http.patch(uri, headers: await _prepareHeaders(headers, auth, uri), body: body, encoding: encoding) : null;
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout), onTimeout: _responseTimeoutHandler) : response;
       } catch (e) {
         Log.d(e?.toString());
@@ -294,7 +293,8 @@ class Network  {
   Future<Http.Response> _delete(url, { Map<String, String> headers, int auth, int timeout }) async {
     if (Connectivity().isNotOffline) {
       try {
-        Future<Http.Response> response = (url != null) ? Http.delete(url, headers: _prepareHeaders(headers, auth, url)) : null;
+        Uri uri = _uriFromUrlString(url);
+        Future<Http.Response> response = (uri != null) ? Http.delete(uri, headers: await _prepareHeaders(headers, auth, uri)) : null;
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout), onTimeout: _responseTimeoutHandler) : response;
       } catch (e) {
         Log.d(e?.toString());
@@ -331,7 +331,8 @@ class Network  {
   Future<String> _read(url, { Map<String, String> headers, int auth, int timeout = 60 }) async {
     if (Connectivity().isNotOffline) {
       try {
-        Future<String> response = (url != null) ? Http.read(url, headers: _prepareHeaders(headers, auth, url)) : null;
+        Uri uri = _uriFromUrlString(url);
+        Future<String> response = (uri != null) ? Http.read(uri, headers: await _prepareHeaders(headers, auth, uri)) : null;
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout)) : response;
       } catch (e) {
         Log.d(e?.toString());
@@ -354,7 +355,8 @@ class Network  {
   Future<Uint8List> _readBytes(url, { Map<String, String> headers, int auth, int timeout = 60 }) async{
     if (Connectivity().isNotOffline) {
       try {
-        Future<Uint8List> response = (url != null) ? Http.readBytes(url, headers: _prepareHeaders(headers, auth, url)) : null;
+        Uri uri = _uriFromUrlString(url);
+        Future<Uint8List> response = (uri != null) ? Http.readBytes(uri, headers: await _prepareHeaders(headers, auth, uri)) : null;
         return ((response != null) && (timeout != null)) ? response.timeout(Duration(seconds: timeout), onTimeout: _responseBytesHandler) : response;
       } catch (e) {
         Log.d(e?.toString());
@@ -374,7 +376,7 @@ class Network  {
     return null;    
   }
 
-  Map<String, String> _prepareHeaders(Map<String, String> headers, int auth, String url) {
+  Future<Map<String, String>> _prepareHeaders(Map<String, String> headers, int auth, Uri uri) async {
 
     Map<String, String> result;
 
@@ -422,8 +424,8 @@ class Network  {
     
 
     //cookies
-    if (url != null) {
-      String cookies = _loadCookiesForRequest(url);
+    if (uri != null) {
+      String cookies = await _loadCookiesForRequest(uri);
       if (AppString.isStringNotEmpty(cookies)) {
         if (result == null) {
           result = (headers != null) ? Map.from(headers) : Map();
@@ -454,7 +456,8 @@ class Network  {
 
 
   void _saveCookiesFromResponse(String url, Http.Response response) {
-    if (AppString.isStringEmpty(url) || response == null)
+    Uri uri = _uriFromUrlString(url);
+    if ((uri == null) || response == null)
       return;
 
     Map<String, String> responseHeaders = response.headers;
@@ -470,31 +473,34 @@ class Network  {
     if (cookiesData == null || cookiesData.length == 0)
       return;
 
-    List<Cookie> cookies = List();
+    List<Cookie> cookies = <Cookie>[];
     for (String cookieData in cookiesData) {
       Cookie cookie = Cookie.fromSetCookieValue(cookieData);
       cookies.add(cookie);
     }
 
     var cj = new CookieJar();
-    cj.saveFromResponse(Uri.parse(url), cookies);
+    cj.saveFromResponse(uri, cookies);
   }
 
-  String _loadCookiesForRequest(String url) {
-    var cj = new CookieJar();
-    List<Cookie> cookies = cj.loadForRequest(Uri.parse(url));
-    if (cookies == null || cookies.length == 0)
-      return null;
+  Future<String> _loadCookiesForRequest(Uri uri) async{
+    if(uri != null) {
+      var cj = new CookieJar();
+      List<Cookie> cookies = await cj.loadForRequest(uri);
+      if (cookies == null || cookies.length == 0)
+        return null;
 
-    String result = "";
-    for (Cookie cookie in cookies) {
-      result += cookie.name + "=" + cookie.value + "; ";
+      String result = "";
+      for (Cookie cookie in cookies) {
+        result += cookie.name + "=" + cookie.value + "; ";
+      }
+
+      //remove the last "; "
+      result = result.substring(0, result.length - 2);
+
+      return result;
     }
-
-    //remove the last "; "
-    result = result.substring(0, result.length - 2);
-
-    return result;
+    return null;
   }
 
   Http.Response _responseTimeoutHandler() {
@@ -503,6 +509,20 @@ class Network  {
 
   Uint8List _responseBytesHandler() {
     return null;
+  }
+
+  Uri _uriFromUrlString(dynamic url){
+    Uri uri;
+    if (url is Uri) {
+      uri = url;
+    }
+    else if (url is String) {
+      uri = Uri.tryParse(url);
+    }
+    else if (url != null) {
+      uri = Uri.tryParse(url.toString());
+    }
+    return uri;
   }
 }