+++ lib/ui/groups/GroupDetailPanel.dart
@@ -285,7 +285,7 @@ class _GroupPanelState extends State<GroupPanel> implements NotificationsListene
   }
 
   Widget _buildGroupInfo() {
-    List<Widget> commands = List<Widget>();
+    List<Widget> commands = [];
 
     String members;
     int membersCount = _group?.membersCount ?? 0;