+++ lib/ui/groups/GroupMembershipRequestPanel.dart
@@ -41,8 +41,8 @@ class _GroupMembershipRequestPanelState extends State<GroupMembershipRequestPane
 
   @override
   void initState() {
-    _focusNodes = List();
-    _controllers = List();
+    _focusNodes = [];
+    _controllers = [];
     _questions = widget.group?.questions ?? [];
     for (int index = 0; index < _questions.length; index++) {
       _controllers.add(TextEditingController());
@@ -173,7 +173,7 @@ class _GroupMembershipRequestPanelState extends State<GroupMembershipRequestPane
 
   void _onSubmit() {
     if (_submitting != true) {
-      List<GroupMembershipAnswer> answers = List<GroupMembershipAnswer>();
+      List<GroupMembershipAnswer> answers = [];
       for (int index = 0; index < _questions.length; index++) {
         String question = _questions[index].question;
         TextEditingController controller = _controllers[index];
@@ -201,7 +201,7 @@ class _GroupMembershipRequestPanelState extends State<GroupMembershipRequestPane
           });
           Navigator.pop(context);
         }
-      }).catchError((){
+      }).catchError((_){
         AppAlert.showDialogResult(context, Localization().getStringEx("panel.membership_request.label.fail", 'Failed to submit request'));
       });
     }