+++ lib/ui/health/HealthStatusPanel.dart
@@ -17,6 +17,7 @@
 import 'dart:async';
 import 'dart:typed_data';
 import 'package:illinois/service/Config.dart';
+import 'package:illinois/service/FlexUI.dart';
 import 'package:intl/intl.dart';
 
 import 'package:flutter/cupertino.dart';
@@ -363,7 +364,7 @@ class _HealthStatusPanelState extends State<HealthStatusPanel> implements Notifi
   List <DropdownMenuItem> _buildCountyDropdownItems(){
     List <DropdownMenuItem> result;
     if (_counties?.isNotEmpty ?? false) {
-      result = List <DropdownMenuItem>();
+      result = [];
       for (HealthCounty county in _counties) {
         result.add(DropdownMenuItem<dynamic>(
           value: county,
@@ -456,7 +457,7 @@ class _HealthStatusPanelState extends State<HealthStatusPanel> implements Notifi
             ),
           ),
         ),
-        Visibility(visible: Health().isVaccinated, child:
+        Visibility(visible: Health().isVaccinated && FlexUI().hasFeature('vaccination_badge'), child:
           Container(width: screenWidth, height: _photoSize, child:
             Align(alignment: Alignment.bottomRight, child:
               Padding(padding: EdgeInsets.only(right: vaccinatedPaddingWidth), child: