+++ lib/ui/onboarding/OnboardingAuthLocationPanel.dart
@@ -202,7 +202,7 @@ class OnboardingAuthLocationPanel extends StatelessWidget with OnboardingPanel {
             Row(
               mainAxisAlignment: MainAxisAlignment.end,
               children: <Widget>[
-                FlatButton(
+                TextButton(
                     onPressed: () {
                       Analytics.instance.logAlert(text: message, selection:okTitle);
                       if (pushNext) {