+++ lib/ui/settings/SettingsHomePanel.dart
@@ -86,6 +86,8 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   @override
   void initState() {
 
+    super.initState();
+
     NotificationService().subscribe(this, [
       Auth.notifyUserPiiDataChanged,
       UserProfile.notifyProfileUpdated,
@@ -98,7 +100,6 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     _loadVersionInfo();
     _refreshHealthUser();
 
-    super.initState();
   }
 
   @override
@@ -253,7 +254,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // Connect
 
   Widget _buildConnect() {
-    List<Widget> contentList = new List();
+    List<Widget> contentList = [];
     contentList.add(Padding(
         padding: EdgeInsets.only(left: 8, right: 8, top: 12, bottom: 2),
         child: Text(
@@ -354,7 +355,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // Customizations
 
   Widget _buildCustomizations() {
-    List<Widget> customizationOptions = new List();
+    List<Widget> customizationOptions = [];
     List<dynamic> codes = FlexUI()['settings.customizations'] ?? [];
     for (int index = 0; index < codes.length; index++) {
       String code = codes[index];
@@ -389,7 +390,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // Connected
 
   Widget _buildConnected() {
-    List<Widget> contentList = new List();
+    List<Widget> contentList = [];
 
     List<dynamic> codes = FlexUI()['settings.connected'] ?? [];
     for (String code in codes) {
@@ -409,7 +410,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   }
 
   List<Widget> _buildConnectedNetIdLayout() {
-    List<Widget> contentList = List();
+    List<Widget> contentList = [];
 
     List<dynamic> codes = FlexUI()['settings.connected.netid'] ?? [];
     for (int index = 0; index < codes.length; index++) {
@@ -452,7 +453,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   }
 
   List<Widget> _buildConnectedPhoneLayout() {
-    List<Widget> contentList = List();
+    List<Widget> contentList = [];
 
     String fullName = Auth().fullUserName ?? '';
     bool hasFullName = AppString.isStringNotEmpty(fullName);
@@ -524,14 +525,14 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
             Row(
               mainAxisAlignment: MainAxisAlignment.end,
               children: <Widget>[
-                FlatButton(
+                TextButton(
                     onPressed: () {
                       Analytics.instance.logAlert(text: "Sign out", selection: "Yes");
                       Navigator.pop(context);
                       Auth().logout();
                     },
                     child: Text(Localization().getStringEx("panel.settings.home.logout.button.yes", "Yes"))),
-                FlatButton(
+                TextButton(
                     onPressed: () {
                       Analytics.instance.logAlert(text: "Sign out", selection: "No");
                       Navigator.pop(context);
@@ -548,7 +549,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // NotificationsOptions
 
   Widget _buildNotifications() {
-    List<Widget> contentList = new List();
+    List<Widget> contentList = [];
 
     List<dynamic> codes = FlexUI()['settings.notifications'] ?? [];
     for (int index = 0; index < codes.length; index++) {
@@ -648,7 +649,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   
 
   Widget _buildCovid19Settings() {
-    List<Widget> contentList = List();
+    List<Widget> contentList = [];
 
     if (Auth().isLoggedIn) {
       if ((_refreshingHealthUser == true) || Health().refreshingUser) {
@@ -702,7 +703,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
             contentList.add(ToggleRibbonButton(
                 height: null,
                 borderRadius: borderRadius,
-                label: Localization().getStringEx("panel.settings.home.covid19.provider_test_result", "Health Provider Test Results"),
+                label: Localization().getStringEx("panel.settings.home.covid19.provider_test_result", "Health Provider Test Results and Vaccine Information"),
                 toggled: (Health().user?.consent == true),
                 context: context,
                 onTap: _onProviderTestResult));
@@ -931,7 +932,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
               return AlertDialog(
                 content: Text(message, style: TextStyle(color: Styles().colors.fillColorPrimary, fontSize: 16, fontFamily: Styles().fontFamilies.bold)),
                 actions: <Widget>[
-                  FlatButton(
+                  TextButton(
                       child: Text(
                           Localization().getStringEx("dialog.yes.title", "Yes"), style: TextStyle(color: Styles().colors.fillColorPrimary, fontSize: 16, fontFamily: Styles().fontFamilies.bold)),
                       onPressed: () {
@@ -939,7 +940,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
                         Navigator.pop(buildContext, true);
                       }
                   ),
-                  FlatButton(
+                  TextButton(
                       child: Text(Localization().getStringEx("dialog.no.title", "No"), style: TextStyle(color: Styles().colors.fillColorPrimary, fontSize: 16, fontFamily: Styles().fontFamilies.bold)),
                       onPressed: () {
                         Analytics.instance.logAlert(text: message, selection: "No");
@@ -979,12 +980,12 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
     Analytics.instance.logSelect(target: "Scan COVID-19 Secret QRcode");
     BarcodeScanner.scan().then((result) {
       // barcode_scan plugin returns 8 digits when it cannot read the qr code. Prevent it from storing such values
-      if (AppString.isStringEmpty(result) || (result.length <= 8)) {
+      if (AppString.isStringEmpty(result?.rawContent) || ((result?.rawContent?.length ?? 0) <= 8)) {
         AppAlert.showDialogResult(context, Localization().getStringEx('panel.settings.home.covid19.alert.qr_code.scan.failed.msg', 'Failed to read QR code.'));
       }
       else {
         setState(() { _scanningHealthUserKeys = true; });
-        _onCovid19QrCodeScanSucceeded(result);
+        _onCovid19QrCodeScanSucceeded(result?.rawContent);
       }
     });
   }
@@ -1030,7 +1031,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // Privacy
 
   Widget _buildPrivacy() {
-    List<Widget> contentList = new List();
+    List<Widget> contentList = [];
 
     List<dynamic> codes = FlexUI()['settings.privacy'] ?? [];
     for (int index = 0; index < codes.length; index++) {
@@ -1066,7 +1067,7 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   // Account
 
   Widget _buildAccount() {
-    List<Widget> contentList = new List();
+    List<Widget> contentList = [];
 
     List<dynamic> codes = FlexUI()['settings.account'] ?? [];
     for (int index = 0; index < codes.length; index++) {
@@ -1246,9 +1247,11 @@ class _SettingsHomePanelState extends State<SettingsHomePanel> implements Notifi
   }
 
   void _updateState() {
-    if (mounted) {
-      setState(() {});
-    }
+    Future.delayed(Duration(), () {
+      if (mounted) {
+        setState(() {});
+      }
+    });
   }
 
 }